move channel to io.parapet
