+++ parapet-testkit/src/main/scala/io/parapet/tests/intg/ChannelSpec.scala
@@ -2,11 +2,10 @@ package io.parapet.tests.intg
 
 import io.parapet.core.Dsl.DslF
 import io.parapet.core.Events.{Start, Stop}
-import io.parapet.{ParConfig, Process}
+import io.parapet.{Channel, ParConfig, Process}
 import io.parapet.runtime.Scheduler.SchedulerConfig
 import io.parapet.core.exceptions.EventHandlingException
-import io.parapet.core.Channel
-import io.parapet.core.Channel.{
+import Channel.{
   ChannelInterruptedException,
   ChannelTimeoutException,
   IllegalChannelStateException,
